[CHORE] constants 분리 및 HELP_MSG 추가
